Fixes #353 Upgrade Dependencies (#360)
Fixes #353 - Upgrade Dependencies
* We resolve a number of compiler warnings.
* Migrated to use lang FieldUtils instead of deprecated Mockito Whitebox for reading private fields.

+++ java-manta-client-unshaded/src/main/java/com/joyent/manta/client/crypto/AesCbcCipherDetails.java
@@ -103,7 +103,7 @@ public final class AesCbcCipherDetails extends AbstractAesCipherDetails {
         final boolean hasIV) {
 
         if (plaintextSize <= 0) {
-            return blockBytes + tagOrHmacBytes;
+            return (long)blockBytes + tagOrHmacBytes;
         }
 
         long calculatedContentLength = plaintextSize;
